[feature/get-suggested-stocks] 추천된 주식 가져오기

+++ market-capitalization/src/main/java/com/shinhansec/marketcapitalization/meeting/dto/MeetingDetailResDto.java
@@ -26,7 +26,7 @@ public class MeetingDetailResDto {
 
     private PurposeType purposeType;
 
-    private int totalInvestment;
+    private long totalInvestment;
 
     private int profitTarget;
